build: CMake presets + windeployqt auto-deploy + Qt auto-detection
- Add CMakePresets.json — cross-platform debug/release presets
- CMakeLists.txt: auto-detect Qt at E:/Qt/{release,debug}/bin
- Post-build: run windeployqt to deploy Qt DLLs next to exe
- Fix qHash return type (uint→size_t, silences MSVC C4267)
- Remove hardcoded E:/Qt/bin path
- Remove platform-specific .bat/.ps1 build scripts

Usage:
  cmake --preset debug    → build_debug/
  cmake --preset release  → build_release/
  cmake --build --preset debug

+++ core/data_cache.h
@@ -15,7 +15,7 @@ struct DataKey {
     }
 };
 
-inline uint qHash(const DataKey &k, uint seed = 0) {
+inline size_t qHash(const DataKey &k, size_t seed = 0) {
     return qHash(k.deviceAddr, seed) ^ qHash(k.registerAddr, seed);
 }
